written setupMocks method
