+++ test/flutter_map_test.dart
@@ -1,81 +1,30 @@
-import 'dart:async';
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:latlong2/latlong.dart';
-import 'package:mockito/mockito.dart';
 
+import 'test_utils/mocks.dart';
 import 'test_utils/test_app.dart';
 
-class MockHttpClientResponse extends Mock implements HttpClientResponse {
-  final _stream = readFile();
-
-  @override
-  int get statusCode => HttpStatus.ok;
-
-  @override
-  int get contentLength => File('test/res/map.png').lengthSync();
-
-  @override
-  HttpClientResponseCompressionState get compressionState =>
-      HttpClientResponseCompressionState.notCompressed;
-
-  @override
-  StreamSubscription<List<int>> listen(void Function(List<int> event)? onData,
-      {Function? onError, void Function()? onDone, bool? cancelOnError}) {
-    return _stream.listen(
-      onData,
-      onError: onError,
-      onDone: onDone,
-      cancelOnError: cancelOnError,
-    );
-  }
-
-  static Stream<List<int>> readFile() => File('test/res/map.png').openRead();
-}
-
-class MockHttpHeaders extends Mock implements HttpHeaders {}
-
-class MockHttpClientRequest extends Mock implements HttpClientRequest {
-  @override
-  HttpHeaders get headers => MockHttpHeaders();
-
-  @override
-  Future<HttpClientResponse> close() => Future.value(MockHttpClientResponse());
-}
-
-class MockClient extends Mock implements HttpClient {
-  @override
-  Future<HttpClientRequest> getUrl(Uri url) {
-    return Future.value(MockHttpClientRequest());
-  }
-}
-
-class MockHttpOverrides extends HttpOverrides {
-  @override
-  HttpClient createHttpClient(SecurityContext? securityContext) => MockClient();
-}
-
 void main() {
-  final markers = <Marker>[
-    Marker(
-      width: 80,
-      height: 80,
-      point: LatLng(45.5231, -122.6765),
-      builder: (_) => const FlutterLogo(),
-    ),
-    Marker(
-      width: 80,
-      height: 80,
-      point: LatLng(40, -120), // not visible
-      builder: (_) => const FlutterLogo(),
-    ),
-  ];
+  setupMocks();
 
   testWidgets('flutter_map', (tester) async {
-    HttpOverrides.global = MockHttpOverrides();
+    final markers = <Marker>[
+      Marker(
+        width: 80,
+        height: 80,
+        point: LatLng(45.5231, -122.6765),
+        builder: (_) => const FlutterLogo(),
+      ),
+      Marker(
+        width: 80,
+        height: 80,
+        point: LatLng(40, -120), // not visible
+        builder: (_) => const FlutterLogo(),
+      ),
+    ];
+
     await tester.pumpWidget(TestApp(markers: markers));
     expect(find.byType(FlutterMap), findsOneWidget);
     expect(find.byType(TileLayer), findsOneWidget);
